requirements.txt
PyJWT
alembic==1.3.2
aniso8601==8.0.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
coverage==4.4.1
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
flask-marshmallow==0.9.0
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.18.1
marshmallow-sqlalchemy==0.16.0
mccabe==0.6.1
more-itertools==8.0.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pyrsistent==0.15.6
pytest==4.3.0
python-dateutil==2.8.1
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.2
six==1.13.0
SQLAlchemy==1.2.18
virtualenv==16.7.9
Werkzeug==0.16.0
wrapt==1.11.2
